# api/docs/swagger.yaml
definitions:
  models.Cars:
    properties:
      fuel:
        type: string
      id:
        type: integer
      name:
        type: string
      office_id:
        type: integer
      reserved:
        type: boolean
      reserved_by:
        $ref: '#/definitions/models.User'
      transmission:
        type: string
      vendor:
        type: string
    type: object
  models.Locations:
    properties:
      Id:
        type: integer
      active:
        type: boolean
      name:
        type: string
    type: object
  models.User:
    properties:
      name:
        type: string
      nation_id:
        type: string
      phone_number:
        type: string
      surname:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/cars/available:
    get:
      description: Get a list of available cars by options(param)
      parameters:
      - description: Date in the format of yyyy-mm-dd
        in: query
        name: receiver_date
        required: true
        type: string
      - description: Date in the format of yyyy-mm-dd
        in: query
        name: delivery_date
        required: true
        type: string
      - description: Time in the format of hh.mm
        in: query
        name: time_start(hour)
        required: true
        type: string
      - description: Time in the format of hh.mm
        in: query
        name: time_end(hour)
        required: true
        type: string
      - description: Location by id
        in: query
        name: location
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Cars'
            type: array
        "400":
          description: missing required parameters, error occurred while converting data
          schema:
            type: string
        "500":
          description: failed to find data, an error occurred while retrieving data from the database, no active offices available, no data found
          schema:
            type: string
      summary: Available Cars
      tags:
      - cars
  /api/cars/reserve:
    post:
      description: Reserve a car by ID and user information
      parameters:
      - description: Car ID
        in: query
        name: id
        required: true
        type: string
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: successfully reserved car {id}
          schema:
            type: string
        "400":
          description: missing required parameters, error occurred while converting data, empty request body, error occurred while decoding JSON
          schema:
            type: string
        "500":
          description: failed to update data, the car is already reserved
          schema:
            type: string
      summary: Reserve a car
      tags:
      - cars
  /api/cars/showreservedcars:
    get:
      description: Retrieve the list of reserved cars from the database.
      operationId: ShowReservedCars
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Cars'
            type: array
        "500":
          description: failed to find data, an error occurred while retrieving data from the database, no data found
          schema:
            type: string
      summary: Show reserved cars
      tags:
      - cars
  /api/locations/show:
    get:
      description: Retrieve a list of active locations
      operationId: ShowLocations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Locations'
            type: array
        "500":
          description: failed to find data, an error occurred while retrieving data from the database, no data found
          schema:
            type: string
      summary: Show active locations
      tags:
      - locations
swagger: "2.0"